requirements.txt
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-javascript>=0.20.0
tree-sitter-cpp>=0.20.0
tree-sitter-java>=0.20.0
tree-sitter-go>=0.20.0
tree-sitter-rust>=0.20.0
plotly>=5.15.0
pandas>=2.0.0
